requirements.txt
boto3==1.15.1
botocore==1.18.1
jmespath==0.10.0
PyJWT==1.7.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
